% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cortisol1.R
\name{cortisol1}
\alias{cortisol1}
\title{Cortisol ELISA 4PL Standard Curve}
\usage{
cortisol1(od450, auto_print = TRUE, auto_plot = TRUE)
}
\arguments{
\item{od450}{长度为 6 的数值向量，对应固定浓度顺序下的原始 OD450。}

\item{auto_print}{是否自动打印结果，默认 TRUE。}

\item{auto_plot}{是否自动显示图形，默认 TRUE。}
}
\value{
返回一个列表，包括数据、参数、R2、100\%OD、Excel公式、图和模型。
}
\description{
输入 6 个标准品原始 OD450，先转换为 OD\%，再基于 OD\% 拟合 4 参数法（4PL）标准曲线。
适用于吸光度越高、浓度越低的竞争法 ELISA。
}
\details{
固定浓度顺序为：
0, 0.41, 1.10, 4.69, 19.3, 82.8 (nmol/L)
}
